op40-move: extract allocation.py from common.py
